+++ Task05/Convert.c
@@ -1,0 +1,64 @@
+#include "Header.h"
+
+extern int k;
+
+void Convert()
+{
+	//	Наполнение стека
+	int N = 71456441;
+
+	ConvertTwo(N);
+
+	if (!overStack)
+	{
+		PrintStack();
+	}
+	else
+	{
+		printf("Стек был переполнен!\n");
+	}
+}
+
+void ConvertTwo(long N)
+{
+	do
+	{
+		printf("С рекурсией (1) или без нее (0)?'\n Введите '1' или '0'\n");
+	} while (!(scanf("%i", &k)) || k > 1);
+
+	if (k == 0)
+	{
+		system("cls");
+		convertToBin(N);
+	}
+	else if (k == 1)
+	{
+		system("cls");
+		convertToBinRecursion(N);
+
+	}
+	else
+	{
+		system("cls");
+		fprintf(stderr, "Вы ввели неверное значение!\n K может быть равно только или 1, или 0.\n Вы ввели %i\n", k);
+	}
+
+}
+
+long convertToBinRecursion(long n)
+{
+	if (n > 0 && !overStack)
+	{
+		push(n % 2);
+		return convertToBinRecursion(n / 2);
+	}
+}
+
+void convertToBin(long n)
+{
+	while (n > 0 && !overStack)
+	{
+		push(n % 2);
+		n /= 2;
+	}
+}
