+++ Task05/Debug/Task05.log
@@ -1,13 +1,24 @@
-﻿  Main.c
-C:\Users\Евгений\OneDrive\C project\Task05\Task05\Header.h(43,9): warning C4033: "push" должен возвращать значение
-C:\Users\Евгений\OneDrive\C project\Task05\Task05\Header.h(57,9): warning C4033: "pop" должен возвращать значение
-C:\Users\Евгений\OneDrive\C project\Task05\Task05\Header.h(89,1): warning C4996: 'scanf': This function or variable may be unsafe. Consider using scanf_s instead. To disable deprecation, use _CRT_SECURE_NO_WARNINGS. See online help for details.
-D:\Windows Kits\10\Include\10.0.18362.0\ucrt\stdio.h(1274): message : см. объявление "scanf"
-C:\Users\Евгений\OneDrive\C project\Task05\Task05\Main.c(42,15): warning C4013: CheckStr: нет определения; предполагается "extern" с возвратом типа "int"
-C:\Users\Евгений\OneDrive\C project\Task05\Task05\Main.c(50,18): warning C4013: CheckFullStack: нет определения; предполагается "extern" с возвратом типа "int"
-C:\Users\Евгений\OneDrive\C project\Task05\Task05\Main.c(68,15): warning C4013: Inverse: нет определения; предполагается "extern" с возвратом типа "int"
-C:\Users\Евгений\OneDrive\C project\Task05\Task05\Header.h(117): warning C4715: convertToBinRecursion: значение возвращается не при всех путях выполнения
-C:\Users\Евгений\OneDrive\C project\Task05\Task05\Header.h(71): warning C4715: pop: значение возвращается не при всех путях выполнения
-C:\Users\Евгений\OneDrive\C project\Task05\Task05\Header.h(50): warning C4716: push: должна возвращать значение
-C:\Users\Евгений\OneDrive\C project\Task05\Task05\Main.c(76): warning C4716: CheckFullStack: должна возвращать значение
-  Task05.vcxproj -> C:\Users\Евгений\OneDrive\C project\Task05\Debug\Task05.exe
+﻿  BracketSequence.c
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Header.h(11,19): error C2079: "Stack" использует не имеющую определения структуру struct "Stack"
+  Main.c
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Header.h(11,19): error C2079: "Stack" использует не имеющую определения структуру struct "Stack"
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Main.c(18,8): error C2224: выражение слева от ".maxSize" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Main.c(20,8): error C2224: выражение слева от ".head" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Main.c(22,9): warning C4013: Convert: нет определения; предполагается "extern" с возвратом типа "int"
+  Struct.c
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Header.h(11,19): error C2079: "Stack" использует не имеющую определения структуру struct "Stack"
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(23,14): error C2371: Stack: переопределение; различные базовые типы
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Header.h(11): message : см. объявление "Stack"
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(27,12): error C2224: выражение слева от ".size" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(27,26): error C2224: выражение слева от ".maxSize" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(34,20): error C2224: выражение слева от ".head" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(35,8): error C2224: выражение слева от ".head" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(36,8): error C2224: выражение слева от ".size" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(41,12): error C2224: выражение слева от ".size" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(50,16): error C2224: выражение слева от ".head" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(51,15): error C2224: выражение слева от ".head" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(52,8): error C2224: выражение слева от ".head" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(52,21): error C2224: выражение слева от ".head" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(56,8): error C2224: выражение слева от ".size" должно иметь тип структуры или объединения
+C:\Users\Евгений\OneDrive\C project\Task05\Task05\Struct.c(62,24): error C2224: выражение слева от ".head" должно иметь тип структуры или объединения
+  Создание кода...
