+++ Task05/Main.c
@@ -14,78 +14,13 @@ int main(int argc, char* argv[])
 {
 	setlocale(LC_ALL, "RUS");
 	//	Максимальный размер стека
+
 	Stack.maxSize = 10;
 	//	Голова обнуляется 
 	Stack.head = NULL;
 	
-#pragma region MyRegion1
-	//	Наполнение стека
-	/*int N = 71456441168230;
-	ConvertTwo(N);
-	if (!overStack)
-	{
-		PrintStack();
-	}
-	else
-	{
-		printf("Стек был переполнен!\n");
-	}*/
-#pragma endregion
+	Convert();
 
-#pragma region MyRegion3
-	char str[] = "(2+(2*){}2))", *p;
-	char sym;
 
-	for (p = str; *p; p++)
-	{
-		sym = *p;
-		if (CheckStr(sym))
-		{
-			//printf("Well done! - %c", sym);
-			if (Stack.head == NULL)
-			{
-				//	Вводим первое значение, если до этого их не было.
-				push(sym);
-			}
-			CheckFullStack(sym);
-		}
-	}
-	if (Stack.head == NULL)
-	{
-		printf("Последовательность правильная!\n");
-	}
-	else
-	{
-		printf("Последовательность не верна! Вот то что осталось!\n");
-		PrintStack();
-	}
-
-#pragma endregion
-
-	return 0;
-}int CheckFullStack(char sym){	Node* begin = Stack.head;	while (begin != NULL)
-	{
-		if (!Inverse(sym))
-		{
-			push(sym);
-		}
-		else
-		{
-			pop();
-		}
-	}}int Inverse(char s){	if ( (Stack.head->value == '(' && s == ')') || (Stack.head->value == '[' && s == ']') || (Stack.head->value == '{' && s == '}') )
-	{
-		return 1;
-	}	else
-	{
-		return 0;
-	}}int CheckStr(str)
-{
-	char t;
-	t = str;
-	if (t == '(' || t == '[' || t == '{' || t == ')' || t == ']' || t == '}')
-	{
-		return 1;
-	}
 	return 0;
 }
